Refactor ContentStorage into new files
