+++ BlueprintUI/Sources/Element/LayoutStorage.swift
@@ -1,0 +1,123 @@
+import Foundation
+
+/// Content storage that supports layout and multiple children.
+struct LayoutStorage<LayoutType: Layout>: ContentStorage {
+
+    var layout: LayoutType
+    var children: [Child]
+
+    init(layout: LayoutType, children: [Child]) {
+        self.layout = layout
+        self.children = children
+    }
+
+    // MARK: ContentStorage
+
+    var childCount: Int {
+        children.count
+    }
+
+    func measure(
+        in constraint: SizeConstraint,
+        environment: Environment,
+        cache: CacheTree
+    ) -> CGSize {
+        cache.get(constraint) { constraint -> CGSize in
+            Logger.logMeasureStart(
+                object: cache.signpostRef,
+                description: cache.name,
+                constraint: constraint
+            )
+            defer { Logger.logMeasureEnd(object: cache.signpostRef) }
+
+            let layoutItems = self.layoutItems(in: environment, cache: cache)
+            return layout.measure(in: constraint, items: layoutItems)
+        }
+    }
+
+    func performLayout(
+        attributes: LayoutAttributes,
+        environment: Environment,
+        cache: CacheTree
+    ) -> [(identifier: ElementIdentifier, node: LayoutResultNode)] {
+        guard children.isEmpty == false else {
+            return []
+        }
+
+        let layoutItems = layoutItems(in: environment, cache: cache)
+        let childAttributes = layout.layout(size: attributes.bounds.size, items: layoutItems)
+
+        var result: [(identifier: ElementIdentifier, node: LayoutResultNode)] = []
+        result.reserveCapacity(children.count)
+
+        var identifierFactory = ElementIdentifier.Factory(elementCount: children.count)
+
+        for index in 0..<children.count {
+            let currentChildLayoutAttributes = childAttributes[index]
+            let currentChild = children[index]
+            let currentChildCache = cache.subcache(
+                index: index,
+                of: children.count,
+                element: currentChild.element
+            )
+
+            let resultNode = LayoutResultNode(
+                element: currentChild.element,
+                layoutAttributes: currentChildLayoutAttributes,
+                environment: environment,
+                children: currentChild.content.performLayout(
+                    attributes: currentChildLayoutAttributes,
+                    environment: environment,
+                    cache: currentChildCache
+                )
+            )
+
+            let identifier = identifierFactory.nextIdentifier(
+                for: type(of: currentChild.element),
+                key: currentChild.key
+            )
+
+            result.append((identifier: identifier, node: resultNode))
+        }
+
+        return result
+    }
+
+    private func layoutItems(
+        in environment: Environment,
+        cache: CacheTree
+    ) -> [(LayoutType.Traits, Measurable)] {
+
+        /// **Note**: We are intentionally using our `indexedMap(...)` and not `enumerated().map(...)`
+        /// here; because the enumerated version is about 25% slower. Because this
+        /// is an extremely hot codepath; this additional performance matters, so we will
+        /// keep track of the index ourselves.
+
+        children.indexedMap { index, child in
+            let childContent = child.content
+            let childCache = cache.subcache(
+                index: index,
+                of: children.count,
+                element: child.element
+            )
+            let measurable = Measurer { constraint -> CGSize in
+                childContent.measure(
+                    in: constraint,
+                    environment: environment,
+                    cache: childCache
+                )
+            }
+
+            return (child.traits, measurable)
+        }
+    }
+
+    struct Child {
+
+        var traits: LayoutType.Traits
+        var key: AnyHashable?
+        var content: ElementContent
+        var element: Element
+
+    }
+}
