+++ BlueprintUI/Sources/Element/LazyStorage.swift
@@ -1,0 +1,61 @@
+import CoreGraphics
+
+/// Content storage that defers creation of its child until measurement or layout time.
+struct LazyStorage: ContentStorage {
+    let childCount = 1
+
+    var builder: (ElementContent.LayoutPhase, SizeConstraint, Environment) -> Element
+
+    func performLayout(
+        attributes: LayoutAttributes,
+        environment: Environment,
+        cache: CacheTree
+    ) -> [(identifier: ElementIdentifier, node: LayoutResultNode)] {
+        let constraint = SizeConstraint(attributes.bounds.size)
+        let child = buildChild(for: .layout, in: constraint, environment: environment)
+        let childAttributes = LayoutAttributes(size: attributes.bounds.size)
+
+        let identifier = ElementIdentifier(elementType: type(of: child), key: nil, count: 1)
+
+        let node = LayoutResultNode(
+            element: child,
+            layoutAttributes: childAttributes,
+            environment: environment,
+            children: child.content.performLayout(
+                attributes: childAttributes,
+                environment: environment,
+                cache: cache.subcache(element: child)
+            )
+        )
+
+        return [(identifier, node)]
+    }
+
+    func measure(in constraint: SizeConstraint, environment: Environment, cache: CacheTree) -> CGSize {
+        cache.get(constraint) { constraint -> CGSize in
+            Logger.logMeasureStart(
+                object: cache.signpostRef,
+                description: cache.name,
+                constraint: constraint
+            )
+
+            defer { Logger.logMeasureEnd(object: cache.signpostRef) }
+
+            let child = buildChild(for: .measurement, in: constraint, environment: environment)
+
+            return child.content.measure(
+                in: constraint,
+                environment: environment,
+                cache: cache.subcache(element: child)
+            )
+        }
+    }
+
+    private func buildChild(
+        for phase: ElementContent.LayoutPhase,
+        in constraint: SizeConstraint,
+        environment: Environment
+    ) -> Element {
+        builder(phase, constraint, environment)
+    }
+}
