+++ BlueprintUI/Sources/Element/MeasurableStorage.swift
@@ -1,0 +1,31 @@
+import CoreGraphics
+
+/// Content storage for leaf nodes.
+struct MeasurableStorage: ContentStorage {
+
+    let childCount = 0
+
+    let measurer: (SizeConstraint, Environment) -> CGSize
+
+    func performLayout(
+        attributes: LayoutAttributes,
+        environment: Environment,
+        cache: CacheTree
+    ) -> [(identifier: ElementIdentifier, node: LayoutResultNode)] {
+        []
+    }
+
+    func measure(in constraint: SizeConstraint, environment: Environment, cache: CacheTree) -> CGSize {
+        cache.get(constraint) { constraint in
+            Logger.logMeasureStart(
+                object: cache.signpostRef,
+                description: cache.name,
+                constraint: constraint
+            )
+
+            defer { Logger.logMeasureEnd(object: cache.signpostRef) }
+
+            return measurer(constraint, environment)
+        }
+    }
+}
