+++ BlueprintUI/Sources/Internal/Measurer.swift
@@ -1,0 +1,8 @@
+import CoreGraphics
+
+struct Measurer: Measurable {
+    var _measure: (SizeConstraint) -> CGSize
+    func measure(in constraint: SizeConstraint) -> CGSize {
+        _measure(constraint)
+    }
+}
